吴竹松 Api 添加(设置条件)接口

+++ 03.Server/xiangqin/src/main/java/jp/frontierinfo/api/indto/A001002E001Input.java
@@ -1,5 +1,6 @@
 package jp.frontierinfo.api.indto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -31,9 +32,9 @@ public class A001002E001Input extends AbstractInput {
 	
 	private String interest;
 	
-	private double uheight;
+	private BigDecimal uheight;
 	
-	private double uweight;
+	private BigDecimal uweight;
 	
 	private String introduce;
